+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Better error messaging on parsing template errors
 - Infra: basic support of RDS: listing, creation and deletion of databases:  `awless list databases`; `awless create/delete database`
 - Access: create an AWS access key for a user
+- DNS: allow to revert creation/deletion of records
+- [#80](https://github.com/wallix/awless/issues/80) DNS: return the ChangeInfo id when creating/deleting a record
 
 ### Bugfixes